Make models optional, add guards and lazy imports
