+++ toktagger/api/models/__init__.py
@@ -1,3 +1,14 @@
 """Contains code for implemented ML models."""
 
-from toktagger.api.models.disruption import DisruptionCNN as DisruptionCNN
+import importlib.util
+
+
+def models_dependencies_installed() -> bool:
+    return (
+        importlib.util.find_spec("torch") is not None
+        and importlib.util.find_spec("ray") is not None
+    )
+
+
+if models_dependencies_installed():
+    from toktagger.api.models.disruption import DisruptionCNN as DisruptionCNN
