+++ toktagger/api/schemas/projects.py
@@ -5,6 +5,7 @@ from enum import Enum
 
 from toktagger.api.core.data_loaders import LoaderRegistry
 from toktagger.api.schemas import ConfiguredModel
+from toktagger.api.models.base import models_dependencies_installed
 
 
 class Task(str, Enum):
@@ -114,6 +115,8 @@ class ProjectIn(ConfiguredModel):
     @computed_field
     @property
     def model_types(self) -> List[str]:
+        if not models_dependencies_installed():
+            return []
         from toktagger.api.models.base import ModelRegistry
 
         return ModelRegistry.names(Task(self.task))
